# src/GameHook.WebAPI/Controllers/FilesController.cs
using GameHook.Domain.Interfaces;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;

namespace GameHook.WebAPI.Controllers
{
    public class MapperFileModel
    {
        public string Id { get; set; } = string.Empty;
        public string DisplayName { get; set; } = string.Empty;
    }

    [ApiController]
    [Produces("application/json")]
    [Consumes("application/json")]
    [Route("files")]
    public class FilesController : ControllerBase
    {
        public IMapperFilesystemProvider MapperFilesystemProvider { get; }

        public FilesController(IMapperFilesystemProvider mapperFilesystemProvider)
        {
            MapperFilesystemProvider = mapperFilesystemProvider;
        }

        [SwaggerOperation("Returns a list of all mapper files available inside of the /mappers folder.")]
        [HttpGet("mappers")]
        public ActionResult<IEnumerable<MapperFileModel>> GetMapperFiles()
        {
            MapperFilesystemProvider.RefreshMapperFiles();

            return Ok(MapperFilesystemProvider.MapperFiles.Select(x => new MapperFileModel()
            {
                Id = x.Id,
                DisplayName = x.DisplayName
            }));
        }
    }
}